Commit tests/out directory to git
The test suite will fail if the out folder does not exist.
This change makes it easier for a new contributor to clone the project
and see the tests run and pass.

Any files produced by tests into the tests/out directory
will still be ignored, but the directory itself will exist.